backend/requirements.txt
# Backend Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.7.0

# LangChain & LLM
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20

# Vector Store & Embeddings
faiss-cpu
numpy>=1.24.0

# CPU-only PyTorch (much smaller, no CUDA)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers>=2.3.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.27.0
tenacity>=8.2.0

# Supabase (for persistent job storage on Railway)
supabase>=2.0.0
